Further refine upload ordering
Update FileComparator to filter out hidden and checksum files when
calculating roots.

Prior to this change it was possible that `maven-metadata.xml.md5` would
be picked as a root file (if the other candidate filenames were long
enough). This then has a knock on effect of retuning an incorrect upload
order.

Fixes gh-4