Don't use interceptor for basic auth header
Update `HttpArtifactory` to use a custom request factory to add
authentication headers.

Prior to this commit, the `HttpArtifactory` used the
`RestTemplateBuilder.basicAuth` method to add headers. This works by
adding a `BasicAuthenticationInterceptor` to the `RestTemplate`.
Unfortunately, adding any interceptor causes the entire message body
to be read into a byte array. This causes an `OutOfMemoryError` whenever
a large file is uploaded.

Closes gh-27