Avoid using Bintray for dependencies
While docker-compose-junit-jupiter 1.5.0 is available from Maven
Central, the default version, 4.18.1, of its
conjure-java-jackson-serialization dependency is not. To workaround
this, dependency management is used to upgrade to 6.16.0 which is
the closest version that is available on Central at this time.
While it's a jump of two major versions, it seems to work for our
purposes – all of the dependencies can now be resolved and the
Docker Composed-based tests now pass.

Closes gh-90