Don't use `builds.find` AQL when the user is not an admin
Update `HttpArtifactoryBuildRuns` to fallback to using regular REST
calls when the user is not an admin. AQL that uses anything other than
`items.find` will silently return no results if the user doesn't have
admin privileges. We now attempt to detect if a user doesn't have admin
rights and if so use the inefficient regular REST API.

Closes gh-112